+++ ndonnx/_core/_stringimpl.py
@@ -10,8 +10,9 @@ import numpy as np
 import ndonnx as ndx
 import ndonnx._data_types as dtypes
 import ndonnx._opset_extensions as opx
-import ndonnx.additional as nda
 
+from ._coreimpl import CoreOperationsImpl
+from ._interface import OperationsBlock
 from ._nullableimpl import NullableOperationsImpl
 from ._shapeimpl import UniformShapeOperations
 from ._utils import binary_op, validate_core
@@ -20,7 +21,7 @@ if TYPE_CHECKING:
     from ndonnx import Array
 
 
-class StringOperationsImpl(UniformShapeOperations):
+class _StringOperationsImpl(OperationsBlock):
     @validate_core
     def add(self, x, y) -> Array:
         return binary_op(x, y, opx.string_concat)
@@ -53,7 +54,7 @@ class StringOperationsImpl(UniformShapeOperations):
         self, x, dtype: dtypes.CoreType | dtypes.StructType | None = None, device=None
     ):
         if dtype is not None and not isinstance(
-            dtype, (dtypes.CoreType, dtypes._NullableCore)
+            dtype, (dtypes.CoreType, dtypes.NullableCore)
         ):
             raise TypeError("'dtype' must be a CoreType or NullableCoreType")
         if dtype in (None, dtypes.utf8, dtypes.nutf8):
@@ -70,17 +71,6 @@ class StringOperationsImpl(UniformShapeOperations):
         return ndx.zeros_like(x, dtype=dtype, device=device)
 
     @validate_core
-    def make_nullable(self, x, null):
-        if null.dtype != dtypes.bool:
-            raise TypeError("'null' must be a boolean array")
-
-        return ndx.Array._from_fields(
-            dtypes.into_nullable(x.dtype),
-            values=x.copy(),
-            null=ndx.broadcast_to(null, nda.shape(x)),
-        )
-
-    @validate_core
     def where(self, condition, x, y):
         if x.dtype != y.dtype:
             target_dtype = ndx.result_type(x, y)
@@ -89,6 +79,11 @@ class StringOperationsImpl(UniformShapeOperations):
         return super().where(condition, x, y)
 
 
-class NullableStringOperationsImpl(StringOperationsImpl, NullableOperationsImpl):
-    def make_nullable(self, x, null):
-        return NotImplemented
+class StringOperationsImpl(
+    CoreOperationsImpl, _StringOperationsImpl, UniformShapeOperations
+): ...
+
+
+class NullableStringOperationsImpl(
+    NullableOperationsImpl, _StringOperationsImpl, UniformShapeOperations
+): ...